pyproject.toml
[tool.poetry]
name = "nautobot-plugin-chatops-grafana"
version = "0.1.0"
description = "Nautobot Plugin Chatops Grafana"
authors = ["Network to Code, LLC <opensource@networktocode.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-chatops-grafana"
repository = "https://github.com/nautobot/nautobot-plugin-chatops-grafana"
keywords = ["nautobot", "nautobot-plugin"]
include = [
    "LICENSE",
    "README.md",
]
packages = [
    { include = "nautobot_plugin_chatops_grafana" },
]

[tool.poetry.plugins."nautobot.workers"]
"grafana" = "nautobot_plugin_chatops_grafana.worker:grafana"

[tool.poetry.dependencies]
# Used for local development
nautobot = "^1.0.0"
python = "^3.7"
nautobot-chatops = "^1.3.1"
pydantic = "^1.8.2"
schema-enforcer = "^1.1.0"
isodate = "^0.6.0"
diffsync = "^1.3.0"

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coverage = "*"
django-debug-toolbar = "*"
flake8 = "*"
invoke = "*"
mkdocs = "*"
pydocstyle = "*"
pylint = "*"
pylint-django = "*"
yamllint = "*"
prybar = "^1.0.0"

[tool.poetry.extras]
nautobot = ["nautobot"]

[tool.schema_enforcer]
main_directory = "nautobot_plugin_chatops_grafana/schema"
data_file_search_directories = ["./nautobot_plugin_chatops_grafana"]

# Currently only YAML files are validated.
data_file_extensions = [".yaml", ".yml"]

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"
